changing data type of input

+++  Day - 10/1_Calculator_functionsDictionary.py
@@ -26,7 +26,8 @@ operations = {
 
 def calculations():
   print(logo)
-  num1 = int(input("What's the first number: "))
+
+  num1 = float(input("What's the first number: "))
 
   for symbol in operations:
     print(symbol)
@@ -35,7 +36,7 @@ def calculations():
   while to_continue:
 
     operation_symbol = input("Pick an operation: ")
-    num2 = int(input("What's the next number: "))
+    num2 = float(input("What's the next number: "))
     calc_function = operations[operation_symbol]
     answer = calc_function(num1, num2)
     print(f"{num1} {operation_symbol} {num2} = {answer}" )
